backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filetype==1.2.0
Flask==3.1.0
frozendict==2.4.6
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.48
langchain-google-genai==2.1.1
langchain-text-splitters==0.3.7
langsmith==0.3.18
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.4
openai==1.68.2
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pickle-mixin==1.0.2
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
yfinance==0.2.55
zstandard==0.23.0
